+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Internal rework to remove the use generic resource to ease out the migration to Mimir.
 - Update generic resource so we can delete resources if mimir is enabled.
 - Remove unused scrape_timeout inhibition.
+- Send oncall towards grafana instead of prometheus by vendoring some of mimir's code.
 
 ## [4.76.0] - 2024-06-03